// lumin-web-backend-2.0/src/Document/enums/organization.roles.enum.ts
export enum OrganizationRoles {
    ALL = 'organization_all',
    ORGANIZATION_ADMIN = 'organization_admin',
    BILLING_MODERATOR = 'billing_moderator',
    MEMBER = 'member',
}

export enum OrganizationDocumentRoles {
    ALL = 'organization_all',
    OWNER = 'organization_owner',
    EDITOR = 'organization_editor',
    VIEWER = 'organization_viewer',
    SPECTATOR = 'organization_spectator',
    SHARER = 'organization_sharer',
}

export enum OrgTeamDocumentRoles {
    ALL = 'org_team_all',
    OWNER = 'org_team_owner',
    EDITOR = 'org_team_editor',
    VIEWER = 'org_team_viewer',
    SPECTATOR = 'org_team_spectator',
    SHARER = 'org_team_sharer',
}